+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'web-chat-server'
+rootProject.name = 'web-auth-server'
