+++ src/main/java/ru/arkham/webauth/WebAuthServerApplication.java
@@ -1,0 +1,12 @@
+package ru.arkham.webauth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebAuthServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WebAuthServerApplication.class, args);
+    }
+}
