--- src/main/java/ru/arkham/webchat/WebChatServerApplication.java
@@ -1,13 +1,0 @@
-package ru.arkham.webchat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WebChatServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WebChatServerApplication.class, args);
-    }
-
-}
