--- src/main/java/ru/arkham/webchat/controller/request/LoginRequest.java
@@ -1,23 +1,0 @@
-package ru.arkham.webchat.controller.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
-
-/**
- * Тело запроса авторизации.
- */
-@Data
-public class LoginRequest {
-
-    /**
-     * Имя пользователя.
-     */
-    @NotBlank
-    private String name;
-
-    /**
-     * Пароль пользователя.
-     */
-    @NotBlank
-    private String password;
-}
