--- src/main/java/ru/arkham/webchat/exception/UserNotFoundException.java
@@ -1,19 +1,0 @@
-package ru.arkham.webchat.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-/**
- * Исключение, обозначающее ошибку поиска информации о пользователе.
- */
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class UserNotFoundException extends RuntimeException {
-
-    /**
-     * Конструктор.
-     * @param message сообщение.
-     */
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
