+++ src/test/java/ru/arkham/webauth/WebAuthServerApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.arkham.webauth;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebAuthServerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
