--- src/test/java/ru/arkham/webchat/WebChatServerApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.arkham.webchat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WebChatServerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
